Cleaned up significant portion of the code: - Rearranged code in source and header files so it makes more sense. - Renamed some functions, mainly glut event handlers. - Removed unnecessary functions declarations in headers and moved them to source files, so they are not included everywhere if not needed. - Removed unnecessary #includes Application now starts in drawing mode, where the simulation is paused and the structure can be drawn and edited. When the simulation is resumed, it is not possible to draw anything. Added total simulation time, which is paused when the simulation is not being run. Cleaned up the interface. Snapping now works for particles. Changed Vector2d constructor, so it now accepts doubles instead of floats.

+++ Trusses/math.cpp
@@ -8,9 +8,9 @@
 
 #include "math.h"
 
-float random(float range)
+double random(double range)
 {
-    float r = 1.0 * rand() / RAND_MAX;
+    double r = 1.0 * rand() / RAND_MAX;
     return r * 2 * range - range;
 }
